Add robot part to the Rapp API
Robot API will be divided into few modules, e.g. communication,
navigation or vision. Those will be implemented in full or in part
for every robot and deployed as library.

Implementation uses pimpl pattern to make library ABI as stable as
possible and allow running the same RAPP on different robots'
without the need for any change in code.